+++ db/jig/mapper.php
@@ -23,6 +23,8 @@
 namespace DB\Jig;
 
 //! Flat-file DB mapper
+use ReturnTypeWillChange;
+
 class Mapper extends \DB\Cursor {
 
 	protected
@@ -156,7 +158,7 @@ class Mapper extends \DB\Cursor {
 	*	@param $ttl int|array
 	*	@param $log bool
 	**/
-	function find($filter=NULL,array $options=NULL,$ttl=0,$log=TRUE) {
+	function find($filter=NULL,array|NULL $options=NULL,$ttl=0,$log=TRUE) {
 		if (!$options)
 			$options=[];
 		$options+=[
@@ -352,7 +354,7 @@ class Mapper extends \DB\Cursor {
 	*	@param $options array
 	*	@param $ttl int|array
 	**/
-	function count($filter=NULL,array $options=NULL,$ttl=0) {
+	function count($filter=NULL,array|NULL $options=NULL,$ttl=0) {
 		$now=microtime(TRUE);
 		$out=count($this->find($filter,$options,$ttl,FALSE));
 		$this->db->jot('('.sprintf('%.1f',1e3*(microtime(TRUE)-$now)).'ms) '.
@@ -522,7 +524,7 @@ class Mapper extends \DB\Cursor {
 	*	Retrieve external iterator for fields
 	*	@return object
 	**/
-	function getiterator() {
+	#[ReturnTypeWillChange] function getiterator() {
 		return new \ArrayIterator($this->cast());
 	}