+++ db/sql/mapper.php
@@ -23,6 +23,8 @@
 namespace DB\SQL;
 
 //! SQL data mapper
+use ReturnTypeWillChange;
+
 class Mapper extends \DB\Cursor {
 
 	//@{ Error messages
@@ -205,7 +207,7 @@ class Mapper extends \DB\Cursor {
 	*	@param $filter string|array
 	*	@param $options array
 	**/
-	function stringify($fields,$filter=NULL,array $options=NULL) {
+	function stringify($fields,$filter=NULL,array|NULL $options=NULL) {
 		if (!$options)
 			$options=[];
 		$options+=[
@@ -303,7 +305,7 @@ class Mapper extends \DB\Cursor {
 	*	@param $options array
 	*	@param $ttl int|array
 	**/
-	function select($fields,$filter=NULL,array $options=NULL,$ttl=0) {
+	function select($fields,$filter=NULL,array|NULL $options=NULL,$ttl=0) {
 		list($sql,$args)=$this->stringify($fields,$filter,$options);
 		$result=$this->db->exec($sql,$args,$ttl);
 		$out=[];
@@ -329,7 +331,7 @@ class Mapper extends \DB\Cursor {
 	*	@param $options array
 	*	@param $ttl int|array
 	**/
-	function find($filter=NULL,array $options=NULL,$ttl=0) {
+	function find($filter=NULL,array|NULL $options=NULL,$ttl=0) {
 		if (!$options)
 			$options=[];
 		$options+=[
@@ -355,7 +357,7 @@ class Mapper extends \DB\Cursor {
 	*	@param $options array
 	*	@param $ttl int|array
 	**/
-	function count($filter=NULL,array $options=NULL,$ttl=0) {
+	function count($filter=NULL,array|NULL $options=NULL,$ttl=0) {
 		$adhoc=[];
 		// with grouping involved, we need to wrap the actualy query and count the results
 		if ($subquery_mode=($options && !empty($options['group']))) {
@@ -733,7 +735,7 @@ class Mapper extends \DB\Cursor {
 	*	Retrieve external iterator for fields
 	*	@return object
 	**/
-	function getiterator() {
+	#[ReturnTypeWillChange] function getiterator() {
 		return new \ArrayIterator($this->cast());
 	}