+++ db/sql/session.php
@@ -212,7 +212,7 @@ class Session extends Mapper {
 			);
 		if ($key)
 			$fw->$key=$this->_csrf;
-		$this->_agent=isset($headers['User-Agent'])?$headers['User-Agent']:'';
+		$this->_agent= $headers['User-Agent'] ?? '';
 		if (strlen($this->_agent) > 300) {
 			$this->_agent = substr($this->_agent, 0, 300);
 		}